refactor: extract useToast and ToastContext to lib/toast-context
