+++ src/App.tsx
@@ -34,7 +34,7 @@ import { ErrorBoundary } from './components/ErrorBoundary'
 import { HomeDashboard } from './components/HomeDashboard'
 import { useKeyboardShortcuts } from './hooks/useKeyboardShortcuts'
 import { buildCommands } from './lib/commands'
-import { useToast } from './components/Toast'
+import { useToast } from './lib/toast-context'
 import { ProgressBar } from './components/ProgressBar'
 import { WorkPlanCard } from './components/WorkPlanCard'
 import { resolvePresetTasks } from './task-presets'
