+++ src/lib/toast-context.ts
@@ -1,0 +1,20 @@
+import { createContext, useContext } from 'react'
+
+export type ToastType = 'success' | 'error' | 'info'
+
+export interface ToastOptions {
+  type: ToastType
+  message: string
+}
+
+export interface ToastContextValue {
+  toast: (options: ToastOptions) => void
+}
+
+export const ToastContext = createContext<ToastContextValue | null>(null)
+
+export function useToast() {
+  const ctx = useContext(ToastContext)
+  if (!ctx) throw new Error('useToast must be used within a ToastProvider')
+  return ctx
+}
